requirements.txt
python-pptx==1.0.2
httpx==0.27.2
jsonschema==4.23.0
pillow==10.4.0
